Refactor preprocessor macros
